Add Dto, Repository, Services, Controllers

+++ src/main/java/com/example/dronecourier/repository/DroneRepository.java
@@ -1,9 +1,9 @@
 package com.example.dronecourier.repository;
 
-import com.example.dronecourier.model.Drone;
+import com.example.dronecourier.entity.model.Drone;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface DroneRepository extends JpaRepository<Drone, Integer> {
+public interface DroneRepository extends JpaRepository<Drone, Long> {
 }
