+++ src/main/java/com/example/dronecourier/repository/EmployeeRepository.java
@@ -1,9 +1,9 @@
 package com.example.dronecourier.repository;
 
-import com.example.dronecourier.model.Employee;
+import com.example.dronecourier.entity.model.Employee;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface EmployeeRepository extends JpaRepository<Employee, Integer> {
+public interface EmployeeRepository extends JpaRepository<Employee, Long> {
 }
