+++ src/main/java/com/example/dronecourier/repository/OrderInDeliveryRepository.java
@@ -1,9 +1,9 @@
 package com.example.dronecourier.repository;
 
-import com.example.dronecourier.model.OrderInDelivery;
+import com.example.dronecourier.entity.model.OrderInDelivery;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface OrderInDeliveryRepository extends JpaRepository<OrderInDelivery, Integer> {
+public interface OrderInDeliveryRepository extends JpaRepository<OrderInDelivery, Long> {
 }
