+++ src/main/java/com/example/dronecourier/repository/OrderItemRepository.java
@@ -1,9 +1,9 @@
 package com.example.dronecourier.repository;
 
-import com.example.dronecourier.model.OrderItem;
+import com.example.dronecourier.entity.model.OrderItem;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface OrderItemRepository extends JpaRepository<OrderItem, Integer> {
+public interface OrderItemRepository extends JpaRepository<OrderItem, Long> {
 }
