+++ src/main/java/com/example/dronecourier/repository/OrderRepository.java
@@ -1,9 +1,9 @@
 package com.example.dronecourier.repository;
 
-import com.example.dronecourier.model.Order;
+import com.example.dronecourier.entity.model.Order;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface OrderRepository extends JpaRepository<Order, Integer> {
+public interface OrderRepository extends JpaRepository<Order, Long> {
 }
